Addition of calculateNumberOfBars() helper. Code refactoring.

+++ src/main/java/com/concrete_calculator/reinforcement/Rebar.java
@@ -16,7 +16,7 @@ public class Rebar {
         return steelType;
     }
 
-    public double getBarDiameter() {
+    public int getBarDiameter() {
         return barDiameter;
     }
 }
